% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-utils.R
\name{analyze}
\alias{analyze}
\title{SimDesign function for analyzing simulated data}
\usage{
analyze(condition, dat, fixed_objects = list(ntest = 10000))
}
\description{
SimDesign function for analyzing simulated data
}
\examples{
condition <- data.frame(n = 100, epv = 0.1, sigma2 = 1, p = 2, rho = 0.9,
prev = 0.1)
dat <- generate(condition)
analyze(condition, dat)
}
