% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-utils.R
\name{auroc}
\alias{auroc}
\title{Compute AUC}
\usage{
auroc(y_true, y_pred)
}
\description{
Compute AUC
}
\examples{
dat <- generateData()
ndat <- generateData()
m <- glm(Y ~ ., data = dat, family = binomial)
preds <- predict(m, type = "response", newdata = ndat)
auroc(ndat$Y, preds)
}
