% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-utils.R
\name{brier}
\alias{brier}
\title{Compute Brier score}
\usage{
brier(y_true, y_pred)
}
\arguments{
\item{y_true}{observed outcome}

\item{y_pred}{predicted outcome}
}
\value{
evaluated loss at predictions and observed outcomes
}
\description{
Compute Brier score
}
\examples{
dat <- generateData()
ndat <- generateData()
m <- glm(Y ~ ., data = dat, family = binomial)
preds <- predict(m, type = "response", newdata = ndat)
brier(ndat$Y, preds)
}
