% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-utils.R
\name{brierVar}
\alias{brierVar}
\title{Compute variance of Brier score contributions}
\usage{
brierVar(y_true, y_pred)
}
\description{
Compute variance of Brier score contributions
}
\examples{
dat <- generateData()
ndat <- generateData()
m <- glm(Y ~ ., data = dat, family = binomial)
preds <- predict(m, type = "response", newdata = ndat)
brierVar(ndat$Y, preds)
}
