% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-utils.R
\name{calibrationInTheLarge}
\alias{calibrationInTheLarge}
\title{Compute calibration in the large}
\usage{
calibrationInTheLarge(y_true, y_pred)
}
\description{
Compute calibration in the large
}
\examples{
dat <- generateData()
ndat <- generateData()
m <- glm(Y ~ ., data = dat, family = binomial)
preds <- predict(m, type = "response", newdata = ndat)
calibrationInTheLarge(ndat$Y, preds)
}
