% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-utils.R
\name{calibrationSlope}
\alias{calibrationSlope}
\title{Compute calibration slope}
\usage{
calibrationSlope(y_true, y_pred)
}
\description{
Compute calibration slope
}
\examples{
dat <- generateData()
ndat <- generateData()
m <- glm(Y ~ ., data = dat, family = binomial)
preds <- predict(m, type = "response", newdata = ndat)
calibrationSlope(ndat$Y, preds)
}
