% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-utils.R
\name{evaluateModel.glmnet}
\alias{evaluateModel.glmnet}
\title{Evaluate glmnet}
\usage{
\method{evaluateModel}{glmnet}(m, newx, y_true, loss, ...)
}
\description{
Evaluate glmnet
}
\examples{
set.seed(10)
dat <- generateData()
ndat <- generateData()
m <- fglmnet(Y ~ ., data = dat, alpha = 0.5, lambda = 0.01, family = "binomial")
evaluateModel(m, newx = as.matrix(ndat[,-1]), y_true = ndat$Y, loss = nll)
}
