% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-utils.R
\name{evaluateModel.ranger}
\alias{evaluateModel.ranger}
\title{Evaluate ranger}
\usage{
\method{evaluateModel}{ranger}(m, newx, y_true, loss, ...)
}
\arguments{
\item{m}{ranger object}

\item{newx}{new covariate data}

\item{y_true}{observed outcome}

\item{loss}{loss function}

\item{...}{additional args to \code{predict.ranger()}}
}
\value{
evaluated loss at predictions and observed outcomes
}
\description{
Evaluate ranger
}
\examples{
dat <- generateData()
ndat <- generateData()
m <- ranger(Y ~ ., data = dat, probability = TRUE)
evaluateModel(m, newx = ndat, y_true = ndat$Y, loss = nll)
}
