% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-utils.R
\name{evaluateModel}
\alias{evaluateModel}
\title{Evaluate model at new observations with \code{loss}}
\usage{
evaluateModel(m, newx, y_true, loss, ...)
}
\description{
Evaluate model at new observations with \code{loss}
}
