% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ainet.R
\name{fglmnet}
\alias{fglmnet}
\title{glmnet with formula interface}
\usage{
fglmnet(formula, data, ...)
}
\arguments{
\item{formula}{formula; model formula}

\item{data}{data.frame; data}

\item{...}{additional args passed to \code{glmnet()}}
}
\value{
object of class \code{"glmnet"}
}
\description{
glmnet with formula interface
}
