% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{filter_coef}
\alias{filter_coef}
\title{Filter coefficients from simulation results}
\usage{
filter_coef(dat, cf = "X.0")
}
\arguments{
\item{dat}{data.frame; simulation results}

\item{cf}{character; names of coefs to filter}
}
\description{
Filter coefficients from simulation results
}
