% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-utils.R
\name{generate}
\alias{generate}
\title{SimDesign function for generating the data}
\usage{
generate(condition, fixed_objects = list(ntest = 10000))
}
\description{
SimDesign function for generating the data
}
\examples{
condition <- data.frame(n = 100, EPV = 20, sparsity = 0.9, prev = 0.01,
sigma2 = 1, rho = 0.9, p = 2, nonlin = TRUE)
generate(condition)
}
