% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-utils.R
\name{generateData}
\alias{generateData}
\title{Generate data from logistic model}
\usage{
generateData(
  n = 100,
  p = 20,
  b = rnorm(p),
  prev = 0.5,
  rho = 0,
  nonlin = FALSE
)
}
\arguments{
\item{n}{integer; sample size}

\item{p}{integer; number of predictors}

\item{b}{numeric; vector of coefficients}

\item{prev}{numeric; outcome prevalence}

\item{rho}{numeric; correlation in predictors}

\item{nonlin}{logical; whether to include a nonlinear effect}
}
\value{
simulated data in data.frame with column "Y" and "X.1" until "X.p"
}
\description{
Generate data from logistic model
}
\examples{
dat <- generateData(p = 5, prev = 0.01, rho = 0.5)
cor(dat[,-1])
}
