% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{read_results}
\alias{read_results}
\title{Read simulation results}
\usage{
read_results(path, which = c("estimands", "coefs"))
}
\arguments{
\item{path}{character; path to simulation results}

\item{which}{character; either \code{"estimands"} or \code{"coefs"}
are extracted from the results}
}
\value{
data.frame/tibble with simulation results
}
\description{
Read simulation results
}
\examples{
res <- read_results("inst/simResults-test/")
ano <- run_anova(data = res)
vis_results(pdat = ano, save = FALSE)
nadat <- res \%>\%
   group_by(n, EPV, prev, rho, sparsity, model) \%>\%
   summarise(frac_na = round(100 * mean(is.na(brier)), 1))
vis_na(nadat, save = FALSE)
vis_calibration(res, save = FALSE)

res <- read_results("inst/simResults-test/", "coefs")
cres <- filter_coef(res, "X.0")
sumr <- summarize_coef(cres)
vis_coefs(cres, save = FALSE)
nadat <- cres \%>\%
   group_by(n, EPV, prev, rho, sparsity, method, coef) \%>\%
   summarise(frac_na = round(100 * mean(is.na(estimate)), 1))
vis_coefs_na(nadat, save = FALSE)

}
