% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{run_anova}
\alias{run_anova}
\title{Analyze simulation results via anova}
\usage{
run_anova(
  formula = brier ~ 0 + fct,
  data,
  conds = with(data, unique(paste0("n", n, "EPV", EPV, "prev", prev, "rho", rho,
    "sparsity", sparsity))),
  models = c("GLM", "EN", "AEN", "RF"),
  compare_against = "AINET"
)
}
\arguments{
\item{formula}{formula; model formula of the form
\code{estimand ~ 0 + experimental_conditions}}

\item{data}{data.frame; contains columns referred to in \code{formula}}

\item{conds}{character; conditions to set up contrasts}

\item{models}{character; which method to compare against
\code{compare_against}}

\item{compare_against}{character; method to compare against}
}
\value{
list of data.frames with anova results (pvalue, contrast, estimates)
}
\description{
Analyze simulation results via anova
}
