% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-utils.R
\name{summarize}
\alias{summarize}
\title{SimDesign function for summarizing simulation results}
\usage{
summarize(condition, results, fixed_objects = NULL)
}
\arguments{
\item{condition}{simulation condition}

\item{results}{results obtained from \code{analyze()}}

\item{fixed_objects}{fixed object}
}
\value{
list of summarized estimates of prediction losses and coefs
}
\description{
SimDesign function for summarizing simulation results
}
\examples{
set.seed(123)
condition <- data.frame(n = 100, epv = 10, sigma2 = 1, p = 10, rho = 0,
prev = 0.5)
dat <- generate(condition)
res <- analyze(condition, dat)
summarize(condition, res)
}
