% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{vis_calibration}
\alias{vis_calibration}
\title{Viz calibration from simulation results}
\usage{
vis_calibration(
  pdat,
  metric = c("cslope", "clarge"),
  save = TRUE,
  lim = c(-100, 100),
  only_one = FALSE
)
}
\description{
Viz calibration from simulation results
}
