% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{vis_coefs}
\alias{vis_coefs}
\title{Viz coefficients from simulation results}
\usage{
vis_coefs(
  pdat,
  cf = "X.0",
  save = TRUE,
  lim = range(pdat$estimate, na.rm = TRUE)
)
}
\description{
Viz coefficients from simulation results
}
