% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{vis_coefs_na}
\alias{vis_coefs_na}
\title{Viz missing coefficients from simulation results}
\usage{
vis_coefs_na(
  pdat,
  rhos = unique(as.numeric(as.character(pdat$rho))),
  sparsities = unique(as.numeric(as.character(pdat$sparsity))),
  cf = "X.0",
  save = TRUE
)
}
\description{
Viz missing coefficients from simulation results
}
