% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{vis_na}
\alias{vis_na}
\title{Visualize missing estimates}
\usage{
vis_na(
  pdat,
  rhos = unique(as.numeric(as.character(pdat$rho))),
  sparsities = unique(as.numeric(as.character(pdat$sparsity))),
  xlab = "brier",
  save = TRUE
)
}
\description{
Visualize missing estimates
}
