% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-utils.R
\name{acc}
\alias{acc}
\title{Compute accuracy}
\usage{
acc(y_true, y_pred)
}
\description{
Compute accuracy
}
\examples{
dat <- generateData()
ndat <- generateData()
m <- glm(Y ~ ., data = dat, family = binomial)
preds <- predict(m, type = "response", newdata = ndat)
acc(ndat$Y, preds)
}
