% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-utils.R
\name{evaluateModel.ranger}
\alias{evaluateModel.ranger}
\title{Evaluate ranger}
\usage{
\method{evaluateModel}{ranger}(m, newx, y_true, loss, ...)
}
\description{
Evaluate ranger
}
\examples{
dat <- generateData()
ndat <- generateData()
m <- ranger(Y ~ ., data = dat, probability = TRUE)
evaluateModel(m, newx = ndat, y_true = ndat$Y, loss = nll)
}
