% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-utils.R
\name{generateData}
\alias{generateData}
\title{Generate data from logistic model}
\usage{
generateData(
  n = 100,
  p = 20,
  b = rnorm(p),
  prev = 0.5,
  rho = 0,
  nonlin = FALSE
)
}
\description{
Generate data from logistic model
}
\examples{
dat <- generateData(p = 5, prev = 0.01, rho = 0.5)
cor(dat[,-1])
}
