% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-utils.R
\name{nll}
\alias{nll}
\title{Compute negative log-likelihood / n (log-score)}
\usage{
nll(y_true, y_pred)
}
\description{
Compute negative log-likelihood / n (log-score)
}
\examples{
dat <- generateData()
ndat <- generateData()
m <- glm(Y ~ ., data = dat, family = binomial)
preds <- predict(m, type = "response", newdata = ndat)
nll(ndat$Y, preds)
}
