% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{run_anova}
\alias{run_anova}
\title{Analyze simulation results via anova}
\usage{
run_anova(
  formula = brier ~ 0 + fct,
  data,
  conds = with(data, unique(paste0("n", n, "EPV", EPV, "prev", prev, "rho", rho,
    "sparsity", sparsity))),
  models = c("GLM", "EN", "AEN", "RF"),
  compare_against = "AINET"
)
}
\description{
Analyze simulation results via anova
}
