% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-utils.R
\name{scaledBrier}
\alias{scaledBrier}
\title{Compute scaled Brier score}
\usage{
scaledBrier(y_true, y_pred)
}
\description{
Compute scaled Brier score
}
\examples{
dat <- generateData()
ndat <- generateData()
m <- glm(Y ~ ., data = dat, family = binomial)
preds <- predict(m, type = "response", newdata = ndat)
scaledBrier(ndat$Y, preds)
}
