% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{vis_results}
\alias{vis_results}
\title{Visualize anova results}
\usage{
vis_results(
  pdat,
  xlab = "brier",
  save = TRUE,
  lim = range(pdat$Estimate, na.rm = TRUE)
)
}
\description{
Visualize anova results
}
